requirements.txt
httpx
python-ranges
